+++ src/global.d.ts
@@ -3,13 +3,7 @@ export {};
 declare global {
   interface Window {
     electronAPI?: {
-      versionChecker: ({
-        account,
-        apps,
-      }: {
-        account: string;
-        apps: string[];
-      }) => Promise<
+      versionChecker: (props: { account: string; apps: string[] }) => Promise<
         [
           {
             account: string;
@@ -18,15 +12,28 @@ declare global {
           },
         ]
       >;
-      updateAccount: (account: string) => Promise<UpdateLog>;
-      manageRelease: (options: {
+      updateAccount: (props: { account: string }) => Promise<UpdateLog>;
+      switchAccount: (props: {
+        account: string;
+      }) => Promise<{ success: boolean; log: string }>;
+      createWorkspace: (props: {
+        account: string;
+        workspace: string;
+      }) => Promise<{ success: boolean; log: string }>;
+      uninstallApps: (props: {
+        account: string;
+        workspace: string;
+        appsToUninstall: string[];
+        forceInstallation?: boolean;
+        forceMaster?: boolean;
+      }) => Promise<{ success: boolean; logs: string[] }>;
+      installApps: (props: {
         account: string;
         workspace: string;
         appsToInstall: string[];
-        appsToUninstall: string[];
-        forceMaster: boolean;
-        stopOnError: boolean;
-      }) => Promise<ReleaseAccountStatus>;
+        forceInstallation?: boolean;
+        forceMaster?: boolean;
+      }) => Promise<{ success: boolean; logs: string[] }>;
     };
   }
 }